Rename nowarn/ to lua/
Also cleanup the includes.
