+++ src/ai/default/aspect_attacks.cpp
@@ -24,7 +24,7 @@
 #include "ai/manager.hpp"
 #include "game_board.hpp"
 #include "log.hpp"
-#include "nowarn/lauxlib.h"
+#include "lua/lauxlib.h"
 #include "map/map.hpp"
 #include "pathfind/pathfind.hpp"
 #include "resources.hpp"
