+++ src/ai/lua/core.cpp
@@ -44,9 +44,7 @@
 #include "ai/composite/contexts.hpp"
 #include "ai/default/aspect_attacks.hpp"
 
-#include "nowarn/lualib.h"
-#include "nowarn/lauxlib.h"
-#include "nowarn/llimits.h"
+#include "lua/lauxlib.h"
 
 static lg::log_domain log_ai_engine_lua("ai/engine/lua");
 #define LOG_LUA LOG_STREAM(info, log_ai_engine_lua)
