+++ src/build_info.cpp
@@ -33,7 +33,7 @@
 #include <fstream>
 #include <iomanip>
 
-#include "nowarn/lua.h"
+#include "lua/lua.h"
 
 #include <SDL2/SDL.h>
 #include <SDL2/SDL_image.h>
