+++ src/scripting/game_lua_kernel.cpp
@@ -122,8 +122,7 @@
 #include <algorithm>
 #include <vector>                       // for vector, etc
 #include <SDL2/SDL_timer.h>                  // for SDL_GetTicks
-#include "nowarn/lauxlib.h"                // for luaL_checkinteger, etc
-#include "nowarn/lua.h"                    // for lua_setfield, etc
+#include "lua/lauxlib.h"                // for luaL_checkinteger, lua_setfield, etc
 
 class CVideo;
