+++ src/scripting/lua_audio.cpp
@@ -15,8 +15,7 @@
 #include "lua_audio.hpp"
 
 #include "log.hpp"
-#include "nowarn/lua.h"
-#include "nowarn/lauxlib.h"
+#include "lua/lauxlib.h"
 #include "scripting/lua_common.hpp"
 #include "scripting/push_check.hpp"
 #include "sound.hpp"
