+++ src/scripting/lua_color.cpp
@@ -16,8 +16,7 @@
 #include "scripting/lua_color.hpp"
 #include "scripting/lua_common.hpp"
 #include "scripting/push_check.hpp"
-#include "nowarn/lauxlib.h"
-#include "nowarn/lua.h"                    // for lua_State, lua_settop, etc
+#include "lua/lauxlib.h"
 #include "log.hpp"
 #include "game_config.hpp"
