+++ src/scripting/lua_gui2.cpp
@@ -49,8 +49,7 @@
 #include <utility>
 #include <vector>
 
-#include "nowarn/lauxlib.h"                // for luaL_checkinteger, etc
-#include "nowarn/lua.h"                    // for lua_setfield, etc
+#include "lua/lauxlib.h"                // for luaL_checkinteger, lua_setfield, etc
 
 static lg::log_domain log_scripting_lua("scripting/lua");
 #define ERR_LUA LOG_STREAM(err, log_scripting_lua)
