+++ src/scripting/lua_rng.cpp
@@ -23,8 +23,7 @@
 #include <new>
 #include <string>
 
-#include "nowarn/lua.h"
-#include "nowarn/lauxlib.h"
+#include "lua/lauxlib.h"
 
 static lg::log_domain log_lua("scripting/lua");
 #define ERR_LUA LOG_STREAM(err, log_lua)
