+++ src/scripting/lua_team.cpp
@@ -26,8 +26,7 @@
 
 #include <string>
 
-#include "nowarn/lua.h"
-#include "nowarn/lauxlib.h"
+#include "lua/lauxlib.h"
 
 /**
  * Implementation for a lua reference to a team,
