+++ src/scripting/lua_terrainmap.cpp
@@ -27,8 +27,7 @@
 #include "game_board.hpp"
 #include "play_controller.hpp"
 
-#include "nowarn/lauxlib.h"
-#include "nowarn/lua.h"
+#include "lua/lauxlib.h"
 
 static lg::log_domain log_scripting_lua("scripting/lua");
 #define LOG_LUA LOG_STREAM(info, log_scripting_lua)
