+++ src/scripting/lua_unit.cpp
@@ -31,8 +31,7 @@
 #include "game_version.hpp"
 #include "deprecation.hpp"
 
-#include "nowarn/lauxlib.h"
-#include "nowarn/lua.h"                    // for lua_State, lua_settop, etc
+#include "lua/lauxlib.h"
 
 static lg::log_domain log_scripting_lua("scripting/lua");
 #define LOG_LUA LOG_STREAM(info, log_scripting_lua)
