+++ src/scripting/lua_widget.cpp
@@ -25,8 +25,7 @@
 #include "scripting/push_check.hpp"
 
 
-#include "nowarn/lauxlib.h"
-#include "nowarn/lua.h"                    // for lua_State, lua_settop, etc
+#include "lua/lauxlib.h"
 
 #include <type_traits>
