+++ src/scripting/lua_widget_methods.cpp
@@ -50,8 +50,7 @@
 #include <utility>
 #include <vector>
 
-#include "nowarn/lauxlib.h"
-#include "nowarn/lua.h"
+#include "lua/lauxlib.h"
 
 static lg::log_domain log_scripting_lua("scripting/lua");
 #define ERR_LUA LOG_STREAM(err, log_scripting_lua)
