+++ src/scripting/mapgen_lua_kernel.cpp
@@ -30,8 +30,7 @@
 #include <string>
 #include <functional>
 
-#include "nowarn/lauxlib.h"
-#include "nowarn/lua.h"
+#include "lua/lauxlib.h"
 #include "scripting/push_check.hpp"
 #include "generators/default_map_generator_job.hpp"
