+++ src/scripting/push_check.hpp
@@ -17,7 +17,7 @@
 #include "scripting/lua_common.hpp"
 #include "scripting/lua_widget.hpp"
 
-#include "nowarn/lauxlib.h"
+#include "lua/lauxlib.h"
 #include "global.hpp"
 #include "tstring.hpp"
 #include "map/location.hpp"
